+++ CHANGELOG.rst
@@ -7,6 +7,15 @@
 Changelog
 =========
 
+2.0.5 - 202X-XX-XX
+------------------
+
+**New feature:**
+
+- Added :meth:`aic`, :meth:`aicc` and :meth:`bic` attributes to the :class:`~glum.GeneralizedLinearRegressorBase`. 
+  These attributes provide the information criteria based on the training data and the effective degrees of freedom 
+  of the maximum likelihood estimate for the model's parameters.
+
 2.0.4 - 202X-XX-XX
 ------------------